split express app & run nodejs server
