+++ server/app.js
@@ -1,16 +1,11 @@
-const startTime = Date.now();
-
-// modules
 const express = require("express");
 const cors = require("cors");
 const bodyParser = require("body-parser");
-const path = require("path");
 
 const routes = require("./routes");
+const database = require("./database");
 
 const app = express();
-
-const database = require("./database");
 
 // utils
 const logger = require("./utils/logger");
@@ -27,58 +22,38 @@ if (!config) {
 // Set the default environment to be `development`
 process.env.NODE_ENV = process.env.NODE_ENV || "development";
 
-database
-	.raw("select 1+1 as result")
+// contains key-value pairs of data submitted in the request body
+app.use(bodyParser.json());
+
+// enable all CORS requests
+app.use(cors());
+
+// import all routes
+app.use(routes);
+
+database.migrate
+	.latest()
 	.then(() => {
-		database.migrate
-			.latest()
-			.then(() => {
-				logger.info("Database migration complete");
-
-				// contains key-value pairs of data submitted in the request body
-				app.use(bodyParser.json());
-
-				// enable all CORS requests
-				app.use(cors());
-
-				// importing all routes modules
-				app.use(routes);
-
-				// Serve vue app
-				if (process.env.NODE_ENV === "production") {
-					app.use(express.static(path.resolve(__dirname, "public")));
-					app.get(/.*/, (req, res) =>
-						res.sendFile(path.resolve(__dirname, "public/index.html"))
-					);
-				}
-
-				// start express server at SERVER_PORT
-				const port = config.server.port || 3000;
-				app.listen(port, () => {
-					logger.info(`LogChimp is running in ${process.env.NODE_ENV}...`);
-					logger.info(`Listening on port: ${port}`);
-					logger.info("Ctrl+C to shut down");
-					logger.info(`LogChimp boot ${(Date.now() - startTime) / 1000}s`);
-				});
-			})
-			.catch(err => {
-				logger.error({
-					code: "DATABASE_MIGRATIONS",
-					message: err
-				});
-			});
+		logger.info("Database migration complete");
 	})
 	.catch(err => {
+		logger.error({
+			code: "DATABASE_MIGRATIONS",
+			message: err
+		});
+
 		if (err.message === "SSL/TLS required") {
 			logger.error({
 				code: "DATABASE_CONNECTION",
 				message: "Enable SSL/TLS on your database connection"
 			});
-		}
 
-		logger.error({
-			code: "DATABASE_CONNECTION",
-			message: `Connecting to database at ${err.address}:${err.port}`,
-			err
-		});
+			logger.error({
+				code: "DATABASE_CONNECTION",
+				message: `Connecting to database at ${err.address}:${err.port}`,
+				err
+			});
+		}
 	});
+
+module.exports = app;
